LPS-61440 move to public package for export
