+++ modules/portal-security/portal-security-sso-token/src/main/java/com/liferay/portal/security/sso/token/constants/TokenConstants.java
@@ -12,14 +12,22 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.token.internal.constants;
+package com.liferay.portal.security.sso.token.constants;
 
 /**
  * @author Mika Koivisto
  */
 public class TokenConstants {
 
+	public static final String AUTH_ENABLED = "enabled";
+
+	public static final String IMPORT_FROM_LDAP = "importFromLDAP";
+
+	public static final String LOGOUT_REDIRECT_URL = "logoutRedirectURL";
+
 	public static final String SERVICE_NAME =
 		"com.liferay.portal.security.sso.token";
+
+	public static final String USER_HEADER = "userTokenName";
 
 }
