+++ modules/portal-security/portal-security-sso-token/src/main/java/com/liferay/portal/security/sso/token/internal/module/configuration/definition/TokenCompanyServiceConfigurationBeanDeclaration.java
@@ -15,7 +15,7 @@
 package com.liferay.portal.security.sso.token.internal.module.configuration.definition;
 
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
-import com.liferay.portal.security.sso.token.internal.module.configuration.TokenConfiguration;
+import com.liferay.portal.security.sso.token.configuration.TokenConfiguration;
 
 import org.osgi.service.component.annotations.Component;
